# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Spector Cognitive Memory & Search API
  description: "Ultra-fast cognitive memory, vector search, and agent orchestration\
    \ platform with biological memory tiers (Working, Episodic, Semantic, Procedural)."
  contact:
    name: Spectrayan Maintainers
    url: https://github.com/spectrayan/spector
    email: support@spectrayan.com
  license:
    name: "Apache License, Version 2.0"
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
- url: /
  description: Current Server Instance
security:
- ApiKeyAuth: []
- BearerAuth: []
...
paths:
...
  /api/v1/config/{category}:
    get:
      tags:
      - config-controller
      operationId: getEffective
      parameters:
      - name: category
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
    put:
      tags:
      - config-controller
      operationId: saveOverride
      parameters:
      - name: category
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
    delete:
      tags:
      - config-controller
      operationId: deleteOverride
      parameters:
      - name: category
        in: path
        required: true
        schema:
          type: string
      - name: scope
        in: query
        required: false
        schema:
          type: string
          default: tenant
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
...
  /api/v1/config/{category}/raw:
    get:
      tags:
      - config-controller
      operationId: getRaw
      parameters:
      - name: category
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
  /api/v1/config/{category}/annotated:
    get:
      tags:
      - config-controller
      operationId: getAnnotated
      parameters:
      - name: category
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  $ref: "#/components/schemas/AnnotatedValue"
  /api/v1/config/schema/{category}:
    get:
      tags:
      - config-controller
      operationId: schema
      parameters:
      - name: category
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
...
components:
  schemas:
...
    AnnotatedValue:
      type: object
      properties:
        value: {}
        source:
          type: string
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      description: Spector API Key Authentication
      name: X-API-Key
      in: header
    BearerAuth:
      type: http
      description: OAuth2 / OIDC JWT Bearer Token
      scheme: bearer
      bearerFormat: JWT