# synapse/spector-synapse/src/main/java/com/spectrayan/spector/synapse/config/api/ConfigController.java
/*
 * Copyright 2026 Spectrayan
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package com.spectrayan.spector.synapse.config.api;

import com.spectrayan.spector.synapse.config.model.ConfigCategory;
import com.spectrayan.spector.synapse.config.model.ScopedConfig;
import com.spectrayan.spector.synapse.config.service.ConfigApplicator;
import com.spectrayan.spector.synapse.config.service.ConfigResolutionService;
import com.spectrayan.spector.synapse.security.SecurityUtils;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.web.bind.annotation.*;

import java.time.Instant;
import java.util.*;

/**
 * REST controller for managing settings configuration overrides in Spector OSS.
 */
@RestController
@RequestMapping("/api/v1/config")
public class ConfigController {

    private static final Logger log = LoggerFactory.getLogger(ConfigController.class);

    private final ConfigResolutionService resolutionService;
    private final ConfigApplicator applicator;
    private final com.spectrayan.spector.synapse.config.service.ConfigSchemaRegistry schemaRegistry;

    public ConfigController(ConfigResolutionService resolutionService,
                            ConfigApplicator applicator,
                            com.spectrayan.spector.synapse.config.service.ConfigSchemaRegistry schemaRegistry) {
        this.resolutionService = resolutionService;
        this.applicator = applicator;
        this.schemaRegistry = schemaRegistry;
    }

    /**
     * Lists all configuration categories.
     */
    @GetMapping("/categories")
    public Map<String, Object> listCategories() {
        var policy = resolutionService.policy();
        var categories = new ArrayList<Map<String, Object>>();

        for (ConfigCategory cat : ConfigCategory.values()) {
            var entry = new LinkedHashMap<String, Object>();
            entry.put("key", cat.key());
            entry.put("displayName", cat.name().replace("_", " "));
            entry.put("tenantOverridable", policy.isTenantOverridable(cat));
            entry.put("userOverridable", policy.isUserOverridable(cat));
            categories.add(entry);
        }

        return Map.of("categories", categories);
    }

    /**
     * Gets the effective configuration for a category.
     */
    @GetMapping("/{category}")
    public Map<String, Object> getEffective(@PathVariable String category) {
        ConfigCategory cat = parseCategory(category);
        String tenantId = SecurityUtils.getTenantId();
        String userId = SecurityUtils.getUserId();

        Map<String, Object> effective = resolutionService.resolve(tenantId, userId, cat);
        return maskApiKeys(effective);
    }

    /**
     * Gets annotated configuration for UI badges.
     */
    @GetMapping("/{category}/annotated")
    public Map<String, ConfigResolutionService.AnnotatedValue> getAnnotated(@PathVariable String category) {
        ConfigCategory cat = parseCategory(category);
        String tenantId = SecurityUtils.getTenantId();
        String userId = SecurityUtils.getUserId();

        Map<String, ConfigResolutionService.AnnotatedValue> annotated =
                resolutionService.resolveAnnotated(tenantId, userId, cat);
        return maskAnnotatedApiKeys(annotated);
    }

    /**
     * Gets raw override configuration at current scope.
     */
    @GetMapping("/{category}/raw")
    public Map<String, Object> getRaw(@PathVariable String category) {
        ConfigCategory cat = parseCategory(category);
        String tenantId = SecurityUtils.getTenantId();
        String userId = SecurityUtils.getUserId();

        Map<String, Object> effective = resolutionService.resolve(tenantId, userId, cat);
        return Map.of("category", cat.key(), "raw", effective);
    }

    /**
     * Saves a configuration override.
     */
    @PutMapping("/{category}")
    @PreAuthorize("hasAnyRole('admin', 'super-admin', 'ADMIN', 'SUPER_ADMIN')")
    public Map<String, Object> saveOverride(@PathVariable String category,
                                            @RequestBody Map<String, Object> request) {
        ConfigCategory cat = parseCategory(category);
        String tenantId = SecurityUtils.getTenantId();
        String userId = SecurityUtils.getUserId();

        String scopeType = (String) request.getOrDefault("scope", "tenant");
        @SuppressWarnings("unchecked")
        Map<String, Object> values = (Map<String, Object>) request.get("values");

        if (values == null || values.isEmpty()) {
            throw new IllegalArgumentException("Missing 'values' in request body");
        }

        ScopedConfig config;
        if ("user".equals(scopeType)) {
            config = ScopedConfig.user(tenantId, userId, cat, values);
        } else {
            config = ScopedConfig.tenant(tenantId, cat, values, userId);
        }

        resolutionService.saveOverride(config);

        // Apply to running system
        Map<String, Object> effective = resolutionService.resolve(tenantId, userId, cat);
        String applyStatus = applicator.apply(tenantId, userId, cat, effective);

        log.info("Config saved: category={}, scope={}, editor={}, status={}", cat.key(), config.scope(), userId, applyStatus);

        return Map.of(
                "status", applyStatus,
                "scope", config.scope(),
                "category", cat.key(),
                "appliedAt", Instant.now().toString()
        );
    }

    /**
     * Deletes a configuration override.
     */
    @DeleteMapping("/{category}")
    @PreAuthorize("hasAnyRole('admin', 'super-admin', 'ADMIN', 'SUPER_ADMIN')")
    public Map<String, Object> deleteOverride(@PathVariable String category,
                                             @RequestParam(defaultValue = "tenant") String scope) {
        ConfigCategory cat = parseCategory(category);
        String tenantId = SecurityUtils.getTenantId();
        String userId = SecurityUtils.getUserId();

        String scopeStr = "user".equals(scope) ? "user:" + tenantId + ":" + userId : "tenant:" + tenantId;
        boolean deleted = resolutionService.removeOverride(scopeStr, cat);

        // Re-apply defaults
        Map<String, Object> effective = resolutionService.resolve(tenantId, userId, cat);
        applicator.apply(tenantId, userId, cat, effective);

        return Map.of(
                "status", deleted ? "deleted" : "not_found",
                "category", cat.key(),
                "scope", scopeStr
        );
    }

    /**
     * Gets the JSON schema fields for a configuration category.
     */
    @GetMapping("/schema/{category}")
    public Map<String, Object> schema(@PathVariable String category) {
        ConfigCategory cat = parseCategory(category);
        List<com.spectrayan.spector.synapse.config.model.ConfigFieldDescriptor> fields =
                schemaRegistry != null ? schemaRegistry.getSchema(cat) : List.of();

        return Map.of(
                "category", cat.key(),
                "fields", fields
        );
    }

    /**
     * Gets the current override policy.
     */
    @GetMapping("/policy")
    public Map<String, Object> policy() {
        return Map.of(
                "overrideOrder", List.of("system", "tenant", "user"),
                "categories", ConfigCategory.values().length
        );
    }

    /**
     * Lists all available providers.
     */
    @GetMapping("/providers/available")
    public Map<String, Object> availableProviders() {
        return Map.of("providers", List.of(
                Map.of("name", "ollama", "type", "local", "status", "available"),
                Map.of("name", "google", "type", "cloud", "status", "available")
        ));
    }

    private ConfigCategory parseCategory(String key) {
        try {
            return ConfigCategory.fromKey(key);
        } catch (IllegalArgumentException e) {
            throw new IllegalArgumentException("Unknown config category: " + key);
        }
    }

    /**
     * Strips any secret-bearing value from a configuration response.
     *
     * <p>Removes rather than masks. Masking was the previous behaviour and it was actively misleading: the
     * response looked handled while the {@code scoped_config} column held the key in cleartext. Secrets are
     * now refused at the write (see {@code ConfigResolutionService.rejectSecretValues}), so a value reaching
     * here is either legacy data or a bug — in both cases the right response is not to serve it at all.</p>
     */
    private Map<String, Object> maskApiKeys(Map<String, Object> values) {
        var sanitised = new LinkedHashMap<>(values);
        ConfigResolutionService.FORBIDDEN_VALUE_KEYS.forEach(sanitised::remove);
        return sanitised;
    }

    /** Annotated-response counterpart of {@link #maskApiKeys(Map)} — removes, does not mask. */
    private Map<String, ConfigResolutionService.AnnotatedValue> maskAnnotatedApiKeys(
            Map<String, ConfigResolutionService.AnnotatedValue> values) {
        var sanitised = new LinkedHashMap<>(values);
        ConfigResolutionService.FORBIDDEN_VALUE_KEYS.forEach(sanitised::remove);
        return sanitised;
    }
}